requirements.txt
# Hungry-Hippo Dependencies
# Emergency Modernism FastAPI + HTMX + Alpine.js

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Templates
jinja2==3.1.3

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database (for future persistence)
sqlalchemy==2.0.25
alembic==1.13.1

# HTTP Client (for external APIs)
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14

# Image Optimization
pillow==10.2.0
pillow-avif-plugin==1.4.1
